requirements.txt
alembic==1.16.5
anyio==4.10.0
arrow==1.3.0
asttokens==3.0.0
binaryornot==0.4.4
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
cookiecutter==2.6.0
cors==1.0.1
debugpy==1.8.16
decorator==5.2.1
executing==2.2.0
filelock==3.19.1
Flask==3.1.2
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flat==0.4
flet==0.28.3
flet-cli==0.28.3
flet-desktop==0.28.3
future==1.0.0
gevent==25.8.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
matplotlib-inline==0.1.7
mdurl==0.1.2
MouseInfo==0.1.3
mysqlclient==2.2.7
nest-asyncio==1.6.0
oauthlib==3.3.1
packaging==25.0
parso==0.8.4
passlib==1.7.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==1.0.9
pyperclip==1.9.0
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytweening==1.2.0
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
qrcode==7.4.2
repath==0.9.0
requests==2.32.5
requests-file==2.1.0
rich==14.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
text-unidecode==1.3
tldextract==5.3.0
toml==0.10.2
tornado==6.5.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
urllib3==2.5.0
watchdog==4.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
zope.event==5.1.1
zope.interface==7.2
